+++ NEWS.md
@@ -1,6 +1,7 @@
 ## NEWS
 
 ### 1.5
+* Added Dia::Sandbox#exit_status for collecting the exit status of a child process running under a sandbox.
 * Dia::Sandbox#run was not returning the PID of the process running under a sandbox like told in the documentation - it is now.
 
 ### 1.4